Create ec2_awslambda.py
This Lambda function works with a CloudWatch Events rule to automatically find running EC2 instances in a specific AWS region. The CloudWatch Events rule is set up to call this lambda function whenever certain events, like creating a new EC2 instance, occur. So, whenever a new instance is created and the instance starts running, the function gets triggered to find and collect the IDs of running instances in that region all this can be viewed in cloudwach logs.